Synced pods method
