+++ internal/cosmos/status_collection_test.go
@@ -1,0 +1,31 @@
+package cosmos
+
+import (
+	"errors"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+	corev1 "k8s.io/api/core/v1"
+)
+
+func TestStatusCollection_SyncedPods(t *testing.T) {
+	var coll StatusCollection
+	require.Empty(t, coll.SyncedPods())
+
+	var catchingUp TendermintStatus
+	catchingUp.Result.SyncInfo.CatchingUp = true
+
+	coll = StatusCollection{
+		{pod: &corev1.Pod{}, status: catchingUp},
+		{pod: &corev1.Pod{}, err: errors.New("some error")},
+	}
+
+	require.Empty(t, coll.SyncedPods())
+
+	var pod corev1.Pod
+	pod.Name = "in-sync"
+	coll = append(coll, PodStatus{pod: &pod})
+
+	require.Len(t, coll.SyncedPods(), 1)
+	require.Equal(t, "in-sync", coll.SyncedPods()[0].Name)
+}
